02_frameworks/2_langchain/code/requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.1
langchain-experimental>=0.0.1

# OpenAI and embeddings
openai>=1.12.0
chromadb>=0.4.22
tiktoken>=0.5.2

# Document processing
unstructured[all-docs]>=0.10.30
pillow>=10.2.0
pydantic>=2.6.1
lxml>=5.1.0
matplotlib>=3.8.0

# Environment and utilities
python-dotenv>=1.0.1 